:construction: WIP render_brackground updade

+++ include/core.h
@@ -6,7 +6,7 @@
 /*   By: cybattis <cybattis@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/12 12:33:14 by cybattis          #+#    #+#             */
-/*   Updated: 2022/06/21 15:29:07 by cybattis         ###   ########.fr       */
+/*   Updated: 2022/06/21 17:23:09 by cybattis         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -136,10 +136,10 @@ typedef struct s_settings
 	int			desired_win_h;
 	int			desired_win_w;
 	double		cam_sensitivity;
-	int64_t		win_slice;
-	int64_t		win_two_slice;
-	int64_t		max_dist;
-	int64_t		max_lerp;
+	double		win_slice;
+	double		win_two_slice;
+	double		max_dist;
+	double		max_lerp;
 }	t_settings;
 
 typedef struct s_light	t_light;
